Split privs commands into modules
